requirements.txt
python-telegram-bot==20.7
PyYAML==6.0.2
pexpect==4.9.0
ansi2html==1.9.1
requests==2.32.3
openai>=1.0.0
pytest==8.3.4
telethon==1.36.0
markdown-it-py==3.0.0
mdit-py-plugins==0.4.1
linkify-it-py==2.0.3
duckduckgo-search==7.5.2
gTTS==2.5.4
youtube-transcript-api==1.2.2
